Fix bot start
