+++ tests/lobby_automation/test_select_brawler.py
@@ -1,6 +1,7 @@
 import unittest
 from unittest.mock import MagicMock, patch
-from PIL import Image
+
+import numpy as np
 
 from lobby_automation import LobbyAutomation
 
@@ -8,22 +9,36 @@ class TestLobbyAutomation(unittest.TestCase):
     
     @patch("lobby_automation.load_toml_as_dict")
     def setUp(self, mock_load_toml):
-        mock_load_toml.return_value = {'lobby': {'select_btn': (0, 0)}}
+        mock_load_toml.return_value = {
+            'lobby': {
+                'brawler_btn': (100, 100),
+                'select_btn': (0, 0),
+            }
+        }
         self.mock_window_controller = MagicMock()
+        self.mock_window_controller.width_ratio = 1
+        self.mock_window_controller.height_ratio = 1
         self.lobby = LobbyAutomation(self.mock_window_controller)
     
     
-    @patch("utils.find_template_center")
-    @patch("stage_manager.load_image")
-    def test_can_select_brawlers(self, mock_load_image, mock_find_template):
+    @patch("lobby_automation.time.sleep", return_value=None)
+    @patch("lobby_automation.extract_text_and_positions")
+    def test_can_select_brawlers(self, mock_extract_text, mock_sleep):
         """Tests that bot can select brawlers once he reaches the brawlers selection menu"""
         EXPECTED_BRAWLER_X = 2012
         EXPECTED_BRAWLER_Y = 978
         TOLERANCE = 50 
         
-        mock_find_template.return_value = (100, 100)
+        mock_extract_text.return_value = {
+            "shelly": {
+                "center": (
+                    EXPECTED_BRAWLER_X / 1.5385,
+                    EXPECTED_BRAWLER_Y / 1.5385,
+                )
+            }
+        }
         
-        test_image = Image.open("./tests/assets/brawlers_menu.PNG")
+        test_image = np.zeros((1080, 1920, 3), dtype=np.uint8)
         self.mock_window_controller.screenshot.return_value = test_image
         
         self.lobby.select_brawler("shelly")
@@ -50,7 +65,7 @@ class TestLobbyAutomation(unittest.TestCase):
             distance_y = abs(actual_y - expected_y)
             
             if distance_x <= tolerance and distance_y <= tolerance:
-                print(f"✓ Click found at ({actual_x}, {actual_y}) within {tolerance}px of ({expected_x}, {expected_y})")
+                print(f"Click found at ({actual_x}, {actual_y}) within {tolerance}px of ({expected_x}, {expected_y})")
                 return True
         
         click_coords = [(call[0][0], call[0][1]) for call in click_calls]
